requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.3
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
